software_team/requirements.txt
torch>=2.6
torchvision>=0.21
streamlit>=1.28
chess>=1.10
Pillow>=10.0
tqdm>=4.60
